add swal dan delete update suggestion

+++ rukun/src/components/VillagerList.jsx
@@ -5,6 +5,8 @@ import { changeAdmin } from '../store/actions/admin'
 import { toast } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
 toast.configure();
+import Swal from 'sweetalert2'
+
 
 export default function VillagerList(props) {
   const { id, name, role, VillageId } = props.user
@@ -12,16 +14,23 @@ export default function VillagerList(props) {
 
   function destroyVillager(id) {
     if(role === 'admin') {
-      toast.error(`cannot delete admin`, {
+      toast.error(`Cannot delete admin`, {
         autoClose: 3000,
         position: toast.POSITION.TOP_RIGHT,
       })
     } else {
-      toast.info(`${name} deleted`, {
-        autoClose: 3000,
-        position: toast.POSITION.TOP_RIGHT,
+      Swal.fire({
+        title: `Are you want to delete this account?, this action is irreverible`,
+        showCancelButton: true,
+        confirmButtonText: `Delete`,
+        denyButtonText: `Cancel`,
+      }).then((result) => {
+        /* Read more about isConfirmed, isDenied below */
+        if (result.isConfirmed) {
+          dispatch(deleteVillagers(id))
+          Swal.fire(`${name} deleted`, '', 'success')
+        }
       })
-      dispatch(deleteVillagers(id))
     }
   }
 
@@ -32,11 +41,18 @@ export default function VillagerList(props) {
         position: toast.POSITION.TOP_RIGHT,
       })
     } else {
-      toast.info(`${name} demoted`, {
-        autoClose: 3000,
-        position: toast.POSITION.TOP_RIGHT,
+      Swal.fire({
+        title: `Are you want to demote this admin?, ensure there is at least one admin or this site will be unaccessable`,
+        showCancelButton: true,
+        confirmButtonText: `Demote`,
+        denyButtonText: `Cancel`,
+      }).then((result) => {
+        /* Read more about isConfirmed, isDenied below */
+        if (result.isConfirmed) {
+          dispatch(changeAdmin(id))
+          Swal.fire('Admin demoted', '', 'success')
+        }
       })
-      dispatch(changeAdmin(id))
     }
   }
